backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-socketio==5.3.6
PyJWT==2.8.0
python-dotenv==1.0.1
vaderSentiment==3.3.2
groq==0.9.0
opencv-python==4.10.0.84
numpy==1.26.4
deepface==0.0.93
mediapipe==0.10.14
fpdf2==2.7.9
eventlet==0.36.1
